requirements.txt
# P360 Project Dependencies
python-pptx>=0.6.21
openpyxl>=3.1.0
PyPDF2>=3.0.0
pandas>=2.0.0
diagrams>=0.24.0
